serve/seggpt_requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
black==24.4.2
captum==0.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@70f454304e1a38378200459dd2dbca0f0f4a5ab4
fairscale==0.4.13
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
grpcio==1.64.1
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
iopath==0.1.9
Jinja2==3.1.4
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numpy==2.0.0
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.0
protobuf==4.25.3
pycocotools==2.0.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.3
six==1.16.0
sympy==1.13.0rc4
setuptools==69.5.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
timm==1.0.7
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yacs==0.1.8
